nbr_pos is now an unsigned long long to catch every int value

+++ includes/structures/node_number_data.h
@@ -6,7 +6,7 @@
 /*   By: user42 <user42@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/15 14:32:57 by user42            #+#    #+#             */
-/*   Updated: 2021/11/16 17:11:46 by user42           ###   ########.fr       */
+/*   Updated: 2021/12/01 15:48:50 by user42           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,8 +15,8 @@
 
 typedef struct	node_number_data
 {
-	int		nbr;
-	int		nbr_pos;
+	int						nbr;
+	unsigned long long		nbr_pos;
 	struct node_number_data	*next;
 }				nbr_data;
